grocery_rec/requirements.txt
Django==4.2.5
pandas==2.0.3
numpy==1.24.4
scikit-surprise==1.1.3
